# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/api/account/login": {
            "post": {
                "description": "登陆",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "account"
                ],
                "summary": "登陆",
                "parameters": [
                    {
                        "description": "登录信息",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.LoginDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "登录失败",
                        "schema": {
                            "$ref": "#/definitions/handler.Response"
                        }
                    }
                }
            }
        },
...
        "/api/attendance_record/approve/query/{leader_staff_id}": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "考勤记录"
                ],
                "summary": "查询主管待审批的考勤记录",
                "parameters": [
                    {
                        "type": "string",
                        "description": "主管员工ID",
                        "name": "leader_staff_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/api/attendance_record/approve_accept/{attendId}": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "考勤记录"
                ],
                "summary": "审批通过考勤记录",
                "parameters": [
                    {
                        "type": "string",
                        "description": "考勤记录ID",
                        "name": "attendId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/api/attendance_record/approve_reject/{attendId}": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "考勤记录"
                ],
                "summary": "审批拒绝考勤记录",
                "parameters": [
                    {
                        "type": "string",
                        "description": "考勤记录ID",
                        "name": "attendId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/api/attendance_record/create": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "考勤记录"
                ],
                "summary": "创建考勤记录",
                "parameters": [
                    {
                        "description": "考勤记录",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.AttendanceRecordCreateDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.Response"
                        }
                    }
                }
            }
        },
        "/api/attendance_record/delete/{attendance_id}": {
            "delete": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "考勤记录"
                ],
                "summary": "删除考勤记录",
                "parameters": [
                    {
                        "type": "string",
                        "description": "考勤记录ID",
                        "name": "attendance_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/api/attendance_record/edit": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "考勤记录"
                ],
                "summary": "更新考勤记录",
                "parameters": [
                    {
                        "description": "更新考勤记录信息",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.AttendanceRecordEditDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.AttendanceRecordEditDTO"
                        }
                    }
                }
            }
        },
        "/api/attendance_record/get_attend_record_is_pay/{staff_id}/{date}": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "考勤记录"
                ],
                "summary": "查询员工指定月份考勤记录是否已结算",
                "parameters": [
                    {
                        "type": "string",
                        "description": "员工ID",
                        "name": "staff_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "日期(格式：YYYY-MM)",
                        "name": "date",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/api/attendance_record/query/{staff_id}": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "考勤记录"
                ],
                "summary": "查询员工考勤记录",
                "parameters": [
                    {
                        "type": "string",
                        "description": "员工ID",
                        "name": "staff_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
...
        "/api/attendance_record/query_history/{staff_id}": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "考勤记录"
                ],
                "summary": "查询员工历史考勤记录",
                "parameters": [
                    {
                        "type": "string",
                        "description": "员工ID",
                        "name": "staff_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/api/authority/create": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "权限管理"
                ],
                "summary": "创建授权",
                "parameters": [
                    {
                        "description": "创建授权",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.AddAuthorityDetailDTO"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/api/authority/edit": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "权限管理"
                ],
                "summary": "编辑授权",
                "parameters": [
                    {
                        "description": "编辑授权",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.UpdateAuthorityDetailDTO"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/api/authority/query_by_user_type/{user_type}": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "权限管理"
                ],
                "summary": "根据用户类型查询授权",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户类型",
                        "name": "user_type",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.GetAuthorityDetailDTO"
                        }
                    }
                }
            }
        },
        "/api/authority/query_by_user_type_and_model": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "权限管理"
                ],
                "summary": "根据用户类型和模块查询授权",
                "parameters": [
                    {
                        "description": "根据用户类型和模块查询授权",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.GetAuthorityDetailDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.GetAuthorityDetailDTO"
                        }
                    }
                }
            }
        },
        "/api/authority/set_admin/{staff_id}": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "权限管理"
                ],
                "summary": "设置管理员",
                "parameters": [
                    {
                        "type": "string",
                        "description": "员工ID",
                        "name": "staff_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/api/authority/set_normal/{staff_id}": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "权限管理"
                ],
                "summary": "设置普通用户",
                "parameters": [
                    {
                        "type": "string",
                        "description": "员工ID",
                        "name": "staff_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
...
        "/api/candidate/accept/{id}": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "候选人管理"
                ],
                "summary": "接受候选人",
                "parameters": [
                    {
                        "type": "string",
                        "description": "候选人ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/api/candidate/create": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "候选人管理"
                ],
                "summary": "创建候选人信息",
                "parameters": [
                    {
                        "description": "候选人信息",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.CandidateCreateDTO"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/api/candidate/delete/{candidate_id}": {
            "delete": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "候选人管理"
                ],
                "summary": "删除候选人信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "候选人ID",
                        "name": "candidate_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/api/candidate/edit": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "候选人管理"
                ],
                "summary": "更新候选人信息",
                "parameters": [
                    {
                        "description": "更新候选人信息",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.CandidateEditDTO"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/api/candidate/query_by_name/{name}": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "候选人管理"
                ],
                "summary": "根据姓名查询候选人",
                "parameters": [
                    {
                        "type": "string",
                        "description": "候选人姓名",
                        "name": "name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页数量",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {}
            }
        },
        "/api/candidate/query_by_staff_id/{staff_id}": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "候选人管理"
                ],
                "summary": "根据员工ID查询候选人",
                "parameters": [
                    {
                        "type": "string",
                        "description": "员工ID",
                        "name": "staff_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页数量",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {}
            }
        },
        "/api/candidate/reject/{id}": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "候选人管理"
                ],
                "summary": "拒绝候选人",
                "parameters": [
                    {
                        "type": "string",
                        "description": "候选人ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
...
        "/api/depart/create": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "部门管理"
                ],
                "summary": "创建部门",
                "parameters": [
                    {
                        "description": "部门信息",
                        "name": "department",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.DepartmentCreateDTO"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/api/depart/del/{dep_id}": {
            "delete": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "部门管理"
                ],
                "summary": "删除部门",
                "parameters": [
                    {
                        "type": "string",
                        "description": "部门ID",
                        "name": "dep_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/api/depart/edit": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "部门管理"
                ],
                "summary": "编辑部门",
                "parameters": [
                    {
                        "description": "部门信息",
                        "name": "department",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.DepartmentEditDTO"
                        }
                    }
                ],
                "responses": {}
            }
        },
...
    },
    "definitions": {
        "handler.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "状态码",
                    "type": "integer"
                },
                "data": {
                    "description": "响应数据"
                },
                "message": {
                    "description": "响应消息"
                },
                "status": {
                    "description": "状态码",
                    "type": "boolean"
                }
            }
        },
        "model.AddAuthorityDetailDTO": {
            "type": "object",
            "required": [
                "authority_content",
                "model",
                "name",
                "user_type"
            ],
            "properties": {
                "authority_content": {
                    "type": "string"
                },
                "model": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "user_type": {
                    "type": "string"
                }
            }
        },
        "model.AttendanceRecordCreateDTO": {
            "type": "object",
            "properties": {
                "date": {
                    "type": "string"
                },
                "leave_days": {
                    "type": "integer"
                },
                "overtime_days": {
                    "type": "integer"
                },
                "staff_id": {
                    "type": "string"
                },
                "staff_name": {
                    "type": "string"
                },
                "work_days": {
                    "type": "integer"
                }
            }
        },
...
        "model.LoginDTO": {
            "type": "object",
            "required": [
                "branch_id",
                "staff_id",
                "user_password"
            ],
            "properties": {
                "branch_id": {
                    "type": "string"
                },
                "staff_id": {
                    "type": "string"
                },
                "user_password": {
                    "type": "string"
                }
            }
        },
...
    }
}